LogDriver: move to subdirectory
